codes/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=4.2.0=py38haa95532_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38h2bbff1b_0
arrow=1.3.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
async-lru=2.0.4=py38haa95532_0
attrs=23.1.0=py38haa95532_0
babel=2.11.0=py38haa95532_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=py38haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.7=py38h9128911_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
ca-certificates=2023.12.12=haa95532_0
certifi=2024.2.2=py38haa95532_0
cffi=1.16.0=py38h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=py38haa95532_0
comm=0.2.1=py38haa95532_0
contourpy=1.0.5=py38h59b6b97_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py38hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
et_xmlfile=1.1.0=py38haa95532_0
exceptiongroup=1.2.0=py38haa95532_0
executing=0.8.3=pyhd3eb1b0_0
filelock=3.13.1=py38haa95532_0
fonttools=4.25.0=pyhd3eb1b0_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=ha860e81_0
fsspec=2023.10.0=py38haa95532_0
gmpy2=2.1.2=py38h7f96b67_0
icu=73.1=h6c2663c_0
idna=3.4=py38haa95532_0
importlib-metadata=7.0.1=py38haa95532_0
importlib_metadata=7.0.1=hd3eb1b0_0
importlib_resources=6.1.1=py38haa95532_1
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.28.0=py38haa95532_0
ipython=8.12.2=py38haa95532_0
ipython-genutils=0.2.0=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.18.1=py38haa95532_1
jinja2=3.1.3=py38haa95532_0
joblib=1.3.2=pypi_0
jpeg=9e=h2bbff1b_1
json5=0.9.6=pyhd3eb1b0_0
jsonpointer=2.4=pypi_0
jsonschema=4.19.2=py38haa95532_0
jsonschema-specifications=2023.7.1=py38haa95532_0
jupyter-contrib-core=0.4.2=pypi_0
jupyter-contrib-nbextensions=0.7.0=pypi_0
jupyter-highlight-selected-word=0.2.0=pypi_0
jupyter-lsp=2.2.0=py38haa95532_0
jupyter-nbextensions-configurator=0.6.3=pypi_0
jupyter_client=8.6.0=py38haa95532_0
jupyter_core=5.5.0=py38haa95532_0
jupyter_events=0.8.0=py38haa95532_0
jupyter_server=2.10.0=py38haa95532_0
jupyter_server_terminals=0.4.4=py38haa95532_1
jupyterlab=4.0.11=py38haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.25.1=py38haa95532_0
kiwisolver=1.4.4=py38hd77b12b_0
krb5=1.20.1=h5b6d351_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_0
libpng=1.6.39=h8cc25b3_0
libpq=12.17=h906ac69_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp-base=1.3.2=h2bbff1b_0
lxml=5.1.0=pypi_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.3=py38h2bbff1b_0
matplotlib=3.7.2=py38haa95532_0
matplotlib-base=3.7.2=py38h4ed8f06_0
matplotlib-inline=0.1.6=py38haa95532_0
mistune=2.0.4=py38haa95532_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py38h2bbff1b_1
mkl_fft=1.3.8=py38h2bbff1b_0
mkl_random=1.2.4=py38h59b6b97_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py38haa95532_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=py_0
nbclient=0.8.0=py38haa95532_0
nbconvert=7.10.0=py38haa95532_0
nbformat=5.9.2=py38haa95532_0
nest-asyncio=1.6.0=py38haa95532_0
networkx=3.1=py38haa95532_0
ninja=1.7.2=0
notebook=7.0.8=py38haa95532_0
notebook-shim=0.2.3=py38haa95532_0
numexpr=2.8.4=py38h7b80656_1
numpy=1.24.3=py38h79a8e48_1
numpy-base=1.24.3=py38h8a87ada_1
openjpeg=2.4.0=h4fc8c34_0
openpyxl=3.0.10=py38h2bbff1b_0
openssl=3.0.13=h2bbff1b_0
overrides=7.4.0=py38haa95532_0
packaging=23.1=py38haa95532_0
pandas=2.0.3=py38h4ed8f06_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=10.2.0=py38h2bbff1b_0
pip=23.3.1=py38haa95532_0
pkgutil-resolve-name=1.3.10=py38haa95532_1
platformdirs=3.10.0=py38haa95532_0
ply=3.11=py38_0
prometheus_client=0.14.1=py38haa95532_0
prompt-toolkit=3.0.43=py38haa95532_0
psutil=5.9.0=py38h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py38haa95532_1
pyparsing=3.0.9=py38haa95532_0
pyqt=5.15.10=py38hd77b12b_0
pyqt5-sip=12.13.0=py38h2bbff1b_0
pyqtwebengine=5.15.6=pypi_0
pyqtwebengine-qt5=5.15.2=pypi_0
python=3.8.18=h1aa4202_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py38haa95532_0
python-json-logger=2.0.7=py38haa95532_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.2.0=cpu_py38hb0bdfb8_0
pytorchtools=0.0.2=pypi_0
pytz=2023.3.post1=py38haa95532_0
pywin32=305=py38h2bbff1b_0
pywinpty=2.0.10=py38h5da7b33_0
pyyaml=6.0.1=py38h2bbff1b_0
pyzmq=25.1.2=py38hd77b12b_0
qt-main=5.15.2=h19c9488_10
referencing=0.30.2=py38haa95532_0
requests=2.31.0=py38haa95532_1
rfc3339-validator=0.1.4=py38haa95532_0
rfc3986-validator=0.1.1=py38haa95532_0
rpds-py=0.10.6=py38h062c2fa_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
send2trash=1.8.2=py38haa95532_0
setuptools=68.2.2=py38haa95532_0
sip=6.7.12=py38hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=py38haa95532_0
soupsieve=2.5=py38haa95532_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.12=py38haa95532_0
tbb=2021.8.0=h59b6b97_0
terminado=0.17.1=py38haa95532_0
threadpoolctl=3.3.0=pypi_0
tinycss2=1.2.1=py38haa95532_0
tomli=2.0.1=py38haa95532_0
tornado=6.3.3=py38h2bbff1b_0
traitlets=5.7.1=py38haa95532_0
types-python-dateutil=2.8.19.20240311=pypi_0
typing-extensions=4.9.0=py38haa95532_1
typing_extensions=4.9.0=py38haa95532_1
uri-template=1.3.0=pypi_0
urllib3=2.1.0=py38haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
webcolors=1.13=pypi_0
webencodings=0.5.1=py38_1
websocket-client=0.58.0=py38haa95532_4
wheel=0.41.2=py38haa95532_0
winpty=0.4.3=4
xz=5.4.6=h8cc25b3_0
yaml=0.2.5=he774522_0
zeromq=4.3.5=hd77b12b_0
zipp=3.17.0=py38haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
